test(backend): add enum validation tests
