+++ backend/src/controllers/mod.rs
@@ -1,2 +1,3 @@
 pub mod seed_controller;
+pub mod seed_controller_tests;
 pub mod variety_controller;
