+++ backend/src/models/dto/new_seed_dto.rs
@@ -9,7 +9,6 @@ use typeshare::typeshare;
 #[typeshare]
 #[derive(Serialize, Deserialize)]
 pub struct NewSeedDTO {
-    pub id: Option<i32>,
     pub name: String,
     pub variety_id: i32,
     pub harvest_year: i16,
